feat : 이체하기 api 생성

+++ server2/server-common/src/main/java/sinhan/server1/domain/account/entity/AccountHistory.java
@@ -31,10 +31,10 @@ public class AccountHistory {
     private LocalDateTime createDate;
 
     @Column(name = "message_code", nullable = true)
-    private Byte messageCode;
+    private int messageCode;
 
     @Builder
-    public AccountHistory(String senderAccountNum, String receiverAccountNum, int amount, LocalDateTime createDate, Byte messageCode) {
+    public AccountHistory(String senderAccountNum, String receiverAccountNum, int amount, LocalDateTime createDate, int messageCode) {
         this.senderAccountNum = senderAccountNum;
         this.receiverAccountNum = receiverAccountNum;
         this.amount = amount;
